Move storage/health endpoint to system API
Relocate storage health check from dives API to system API where it
belongs. This endpoint provides system-level storage status and should
not be tied to dive-specific functionality.

Changes:
- Move /storage/health endpoint from dives_utils.py to system.py
- Update endpoint path from /api/v1/dives/storage/health to /api/v1/admin/system/storage/health
- Update all test references to use new path
- Update test patches to target system router instead of dives router
- Remove unused r2_storage import from dives_utils.py
- Add r2_storage import to system.py

All tests pass and endpoint functionality remains unchanged.
