+++ backend/tests/test_dive_profile_api.py
@@ -104,7 +104,7 @@ class TestDiveProfileAPI:
         test_dive.profile_xml_path = "user_1/2025/09/nonexistent.json"
         db_session.commit()
         
-        with patch('app.routers.dives.dives_utils.r2_storage') as mock_r2:
+        with patch('app.routers.system.r2_storage') as mock_r2:
             mock_r2.download_profile.return_value = None
             
             response = client.get(f"/api/v1/dives/{test_dive.id}/profile", headers=auth_headers)
@@ -180,7 +180,7 @@ class TestDiveProfileAPI:
         test_dive.is_private = True
         test_dive.profile_xml_path = "user_1/2025/09/test_profile.json"
 
-        with patch('app.routers.dives.dives_utils.r2_storage') as mock_r2:
+        with patch('app.routers.system.r2_storage') as mock_r2:
             mock_r2.download_profile.return_value = json.dumps(sample_profile_data).encode('utf-8')
             response = client.get(f"/api/v1/dives/{test_dive.id}/profile")
             assert response.status_code == status.HTTP_403_FORBIDDEN
@@ -278,7 +278,7 @@ class TestDiveProfileAPI:
         """Test successful dive profile deletion."""
         test_dive.profile_xml_path = "user_1/2025/09/test_profile.json"
         
-        with patch('app.routers.dives.dives_utils.r2_storage') as mock_r2:
+        with patch('app.routers.system.r2_storage') as mock_r2:
             mock_r2.delete_profile.return_value = True
             
             response = client.delete(f"/api/v1/dives/{test_dive.id}/profile", 
@@ -318,7 +318,7 @@ class TestDiveProfileAPI:
 
     def test_delete_user_profiles_admin_success(self, client, admin_headers):
         """Test successful deletion of all user profiles by admin."""
-        with patch('app.routers.dives.dives_utils.r2_storage') as mock_r2:
+        with patch('app.routers.system.r2_storage') as mock_r2:
             mock_r2.delete_user_profiles.return_value = True
             
             response = client.delete("/api/v1/dives/profiles/user/1", 
@@ -342,7 +342,7 @@ class TestDiveProfileAPI:
 
     def test_storage_health_check_success(self, client):
         """Test storage health check endpoint."""
-        with patch('app.routers.dives.dives_utils.r2_storage') as mock_r2:
+        with patch('app.routers.system.r2_storage') as mock_r2:
             mock_r2.health_check.return_value = {
                 "r2_available": True,
                 "bucket_accessible": True,
@@ -350,7 +350,7 @@ class TestDiveProfileAPI:
                 "local_storage_writable": True
             }
             
-            response = client.get("/api/v1/dives/storage/health")
+            response = client.get("/api/v1/admin/system/storage/health")
             
             assert response.status_code == status.HTTP_200_OK
             data = response.json()
@@ -361,10 +361,10 @@ class TestDiveProfileAPI:
 
     def test_storage_health_check_error(self, client):
         """Test storage health check endpoint with error."""
-        with patch('app.routers.dives.dives_utils.r2_storage') as mock_r2:
+        with patch('app.routers.system.r2_storage') as mock_r2:
             mock_r2.health_check.side_effect = Exception("Storage error")
             
-            response = client.get("/api/v1/dives/storage/health")
+            response = client.get("/api/v1/admin/system/storage/health")
             
             assert response.status_code == status.HTTP_200_OK
             data = response.json()
